More robust publishing and consuming
Using the recovery test script, I was able to simulate poor network
conditions, which highlighted some areas where the bunny default
recovery wasn't so good. This will ensure that publishing will retry
and reset the connection in the correct cases. This will ensure that
the consumer auto-recovers in case connection is lost.